requirements.txt
asgiref==3.7.2
boto3==1.26.160
botocore==1.29.160
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.2
django-allauth==0.54.0
django-storages==1.13.2
idna==3.4
jmespath==1.0.1
oauthlib==3.2.2
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
stripe==5.4.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
